requirements.txt
# Install: pip install -r requirements.txt

# Core
streamlit>=1.28.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0

# PDF processing (for build_index.py)
PyPDF2>=3.0.0

# Text chunking (for build_index.py)
langchain-text-splitters>=0.3.0

#LangChain RAG path (set USE_LANGCHAIN=true in .env)
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-text-splitters>=0.3.0
